pyproject.toml
[tool.poetry]
name = "mellonpass"
version = "0.1.0"
description = "API For MellonPass UI."
authors = ["Roel Delos Reyes <badcoder28@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2"
strawberry-graphql = "^0.237.3"
django-environ = "^0.11.2"
django-extensions = "^3.2.3"
dj-database-url = "^2.2.0"
psycopg2 = "^2.9.9"
cryptography = "^43.0.1"
pyjwt = "^2.9.0"
pyyaml = "^6.0.2"
ua-parser = "^0.18.0"
user-agents = "^2.2.0"
django-debug = "^0.1.1"
celery = "^5.4.0"
django-celery-results = "^2.5.1"
django-celery-beat = "^2.7.0"
django-ipware = "^7.0.1"
marshmallow = "^3.22.0"
redis = "^5.1.0"
django-ratelimit = "^4.1.0"
django-cors-headers = "^4.4.0"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
mypy = "^1.11.1"
isort = "^5.13.2"
flake8 = "^7.1.1"
django-stubs = "^5.0.4"
ipdb = "^0.13.13"
pytest-django = "^4.8.0"
factory-boy = "^3.3.1"
freezegun = "^1.5.1"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
